# src/main/resources/static/api-docs/api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/v1/pacientes/{id}": {
      "get": {
        "tags": [
          "Paciente"
        ],
        "summary": "Busca um paciente pelo ID",
        "operationId": "buscarPaciente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paciente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paciente"
                }
              }
            }
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Paciente"
        ],
        "summary": "Atualiza um paciente existente",
        "operationId": "atualizarPaciente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Paciente"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Paciente não encontrado"
          },
          "200": {
            "description": "Paciente atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paciente"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos para atualização"
          }
        }
      },
      "delete": {
        "tags": [
          "Paciente"
        ],
        "summary": "Exclui um paciente",
        "operationId": "excluirPaciente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Paciente não encontrado"
          },
          "200": {
            "description": "Paciente excluído com sucesso"
          }
        }
      }
    },
...
    "/v1/pacientes": {
      "post": {
        "tags": [
          "Paciente"
        ],
        "summary": "Cadastra um novo paciente",
        "operationId": "cadastrarPaciente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Paciente"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Dados inválidos para cadastro"
          },
          "201": {
            "description": "Paciente cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paciente"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/pacientes/email/{email}": {
      "get": {
        "tags": [
          "Paciente"
        ],
        "summary": "Busca um paciente pelo email",
        "operationId": "buscarPacientePorEmail",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paciente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paciente"
                }
              }
            }
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        }
      }
    },
    "/v1/pacientes/cpf/{cpf}": {
      "get": {
        "tags": [
          "Paciente"
        ],
        "summary": "Busca um paciente pelo CPF",
        "operationId": "buscarPacientePorCpf",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paciente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paciente"
                }
              }
            }
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Paciente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time"
          },
          "cpf": {
            "type": "string"
          }
        }
      },
...
    }
  }
}